In progress with refactor of param <-> recording
The old relationship between the paramloader, a recording,
and a loader was very unclear.
The new idea is a bit simpler:
A recording contains a paramhandler (or a dict)
and a loader.
The loader populates the recording with any metadata etc
by parsing the dict or paramhandler.
Then at load time, the loader uses this metadata to establish where to load the actual data from